requirements.txt
pandas==0.24.2
numpy==1.16.4
Pillow==6.1.0
dill==0.3.0
torch==1.2.0
torchprof==0.3.1
torchvision==0.4.0
PyYAML==5.1.2
